Fixes #18888 - refactor powerStatus
