+++ webpack/assets/javascripts/react_app/redux/actions/hosts/powerStatus/index.js
@@ -2,8 +2,8 @@ import {
   HOST_POWER_STATUS_REQUEST,
   HOST_POWER_STATUS_SUCCESS,
   HOST_POWER_STATUS_FAILURE
-} from '../../consts';
-import { ajaxRequestAction } from '../common';
+} from '../../../consts';
+import { ajaxRequestAction } from '../../common';
 
 export const getHostPowerState = ({id, url}) => dispatch => ajaxRequestAction({
   dispatch,
