+++ webpack/assets/javascripts/react_app/redux/reducers/hosts/index.js
@@ -1,30 +1,6 @@
-import {
-  HOST_POWER_STATUS_REQUEST,
-  HOST_POWER_STATUS_SUCCESS,
-  HOST_POWER_STATUS_FAILURE
-} from '../../consts';
-import Immutable from 'seamless-immutable';
+import { combineReducers } from 'redux';
+import powerStatus from './powerStatus';
 
-const initialState = Immutable({
-  powerStatus: Immutable({})
+export default combineReducers({
+    powerStatus
 });
-
-export default (state = initialState, action) => {
-  const { payload } = action;
-
-  switch (action.type) {
-    case HOST_POWER_STATUS_REQUEST:
-    case HOST_POWER_STATUS_SUCCESS:
-      return state.setIn(
-        ['powerStatus', payload.id],
-        payload
-      );
-    case HOST_POWER_STATUS_FAILURE:
-      return state.setIn(
-        ['powerStatus', payload.id],
-        { error: payload.error }
-      );
-    default:
-      return state;
-  }
-};
