+++ webpack/assets/javascripts/react_app/redux/reducers/hosts/powerStatus/powerStatus.fixtures.js
@@ -1,7 +1,5 @@
 import Immutable from 'seamless-immutable';
-export const initialState = Immutable({
-  powerStatus: Immutable({})
-});
+export const initialState = Immutable({});
 
 export const request = {
     id: '2',
@@ -16,23 +14,17 @@ export const response = {
 export const error = 'some error happened';
 
 export const stateBeforeResponse = Immutable({
-  powerStatus: {
     [request.id]: request
-  }
 });
 
 export const stateAfterSuccess = Immutable({
-  powerStatus: Immutable({
     [request.id]: {
       ...response
     }
-  })
 });
 
 export const stateAfterFailure = Immutable({
-  powerStatus: Immutable({
     [request.id]: {
       error
     }
-  })
 });
